Refactored helper functions out of voucher test
