+++ backend/smart_contracts/test/voucher.js
@@ -1,35 +1,8 @@
 const Institution = artifacts.require("Institution");
 const truffleAssert = require('truffle-assertions');
+const helper = require("../helpers/truffleTestHelper");
 
 const days = (24*60*60);
-
-
-advanceBlock = () => {
-    return new Promise((resolve, reject) => {
-        web3.currentProvider.send({
-            jsonrpc: "2.0",
-            method: "evm_mine",
-            id: new Date().getTime()
-        }, (err, result) => {
-            if (err) { return reject(err); }
-            const newBlockHash = web3.eth.getBlock('latest').hash;
-            return resolve(newBlockHash)
-        });
-    });
-};
-advanceTime = (time) => {
-    return new Promise((resolve, reject) => {
-        web3.currentProvider.send({
-            jsonrpc: "2.0",
-            method: "evm_increaseTime",
-            params: [time],
-            id: new Date().getTime()
-        }, (err, result) => {
-            if (err) { return reject(err); }
-            return resolve(result);
-        });
-    });
-};
 
 
 contract('Institution', (accounts) => {
@@ -105,8 +78,7 @@ contract('Institution', (accounts) => {
 
         await uut.addVoucher(receiver, web3.utils.fromAscii(test_description), expires_days, {from: owner});
 
-        await advanceTime(1 + expires_days * days); // wait for voucher to expire
-        await advanceBlock();
+        await helper.advanceTimeAndBlock(1 + expires_days * days); // wait for voucher to expire
 
         await truffleAssert.reverts(
             uut.redeem(voucher_index, {from: receiver})
@@ -121,7 +93,8 @@ contract('Institution', (accounts) => {
             uut.redeem(voucher_index, {from: receiver})
         );
     });
-        it('cant redeem non-existant voucher', async () => {
+
+    it('cant redeem non-existant voucher', async () => {
         const receiver = accounts[1];
         await uut.addVoucher(receiver, web3.utils.fromAscii(test_description), 0, {from: owner});
